Updated adventure to new file format
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@1029 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ Adventure/AppBuilder/AppBuilder/CIL.cpp
@@ -25,7 +25,7 @@ unsigned CBNEROW::execute(ExecutionContext& ctx, unsigned pc){
     return pc+mOffset;
   std::vector<Vec2i> breakinfo;
   Vec2i extent = Engine::instance()->getFontRenderer()->getTextExtent(mText, Engine::instance()->getFontID(), breakinfo);
-  extent.y /= breakinfo.size();
+  extent.y /= (int)breakinfo.size();
   if (!Engine::instance()->getInterpreter()->isTSTopToBottom())
     Engine::instance()->getInterpreter()->tsPos().y -= extent.y;
   Vec2i butsize(Engine::instance()->getInterpreter()->getTSWidth(), extent.y);
